requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning - Classical
scikit-learn>=1.3.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# NLP Libraries
spacy>=3.6.0
nltk>=3.8
textblob>=0.17.0
vaderSentiment>=3.3.2

# Visualization
plotly>=5.15.0
wordcloud>=1.9.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Web Deployment
streamlit>=1.25.0
streamlit-drawable-canvas>=0.9.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# For downloading spaCy model (run after install)
# python -m spacy download en_core_web_sm
